server/leads/requirements.txt
Django
djangorestframework
dj-database-url
psycopg2-binary
python-dotenv
pika
gunicorn
scrapy>=2.11
requests>=2.28
setuptools>=65.0  # provides distutils for Python 3.12+
selenium>=4.0
undetected-chromedriver>=3.5